TT#14008 eliminate duplicate packets_lost counters
This solves inconsistent stat output

(cherry picked from commit 785ed90bba8ee8482a7da037fef7adf350bd79be)

+++ daemon/call.c
@@ -3243,7 +3243,7 @@ void call_destroy(struct call *c) {
 			ilog(LOG_INFO, "--- SSRC %s%" PRIx32 "%s", FMT_M(se->h.ssrc));
 			ilog(LOG_INFO, "------ Average MOS %" PRIu64 ".%" PRIu64 ", "
 					"lowest MOS %" PRIu64 ".%" PRIu64 " (at %u:%02u), "
-					"highest MOS %" PRIu64 ".%" PRIu64 " (at %u:%02u) lost:%d",
+					"highest MOS %" PRIu64 ".%" PRIu64 " (at %u:%02u) lost:%u",
 				se->average_mos.mos / mos_samples / 10,
 				se->average_mos.mos / mos_samples % 10,
 				se->lowest_mos->mos / 10,
@@ -3254,7 +3254,7 @@ void call_destroy(struct call *c) {
 				se->highest_mos->mos % 10,
 				(unsigned int) (timeval_diff(&se->highest_mos->reported, &c->created) / 1000000) / 60,
 				(unsigned int) (timeval_diff(&se->highest_mos->reported, &c->created) / 1000000) % 60,
-				se->packets_lost);
+				(unsigned int) se->packets_lost);
 next_k:
 			k = g_list_delete_link(k, k);
 		}
